requirements.txt
phidata
yfinance
python-dotenv
packaging
fastapi
uvicorn
streamlit
groq
duckduckgo-search